% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{country_trans}
\alias{country_trans}
\title{Retrieve a data frame of country abbreviation to name mappings}
\usage{
country_trans()
}
\description{
Provided by <country.io>
}
\note{
Consider caching this information. While our world is, indeed, in turmoil
countries do not disappear \emph{that} quickly.
}
