% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipinfo-package.R
\docType{package}
\name{ipinfo}
\alias{ipinfo}
\alias{ipinfo-package}
\title{Collect Metadata on 'IP' Addresses and Autonomous Systems}
\description{
The 'ipinfo.io' \url{https://ipinfo.io/} service provides an 'API' for
retrieving metadata on 'IP' addresses and autonomous systems. Methods are provided
to query and retrieve this metadata.
}
\author{
Bob Rudis (bob@rud.is)
}
