% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{query_asn}
\alias{query_asn}
\title{Retrieve full ASN details}
\usage{
query_asn(asn, token = Sys.getenv("IPINFOIO_TOKEN", unset = ""))
}
\arguments{
\item{asn}{AS number to query for. The \code{AS} prefix will be provided if not present}

\item{token}{Your <ipinfo.io> access token. A token is  required for the ASN API.}
}
\description{
Retrieve full ASN details
}
\references{
\url{https://ipinfo.io/developers/asn}
}
