% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{query_ip}
\alias{query_ip}
\title{Retrieve full IP details}
\usage{
query_ip(ip, token = Sys.getenv("IPINFOIO_TOKEN", unset = ""))
}
\arguments{
\item{ip}{IP address to query for}

\item{token}{Your <ipinfo.io> access token. A token is not required for the
free tier but is required to gain access to any extra fields that
you subscribed to.}
}
\description{
Retrieve full IP details
}
\note{
Free usage of our API is limited to 1,000 API requests per day.
}
\references{
\url{https://ipinfo.io/developers#full-ip-details}
}
